Set Babel; Refactoring resolvers
