+++ server/src/index.js
@@ -1,54 +1,6 @@
-const { prisma } = require("./generated/prisma-client");
-const { GraphQLServer } = require("graphql-yoga");
-
-const resolvers = {
-  Query: {
-    streams(root, args, context) {
-      return context.prisma.streams();
-    },
-    stream(root, args, context) {
-      return context.prisma.stream({ id: args.streamId });
-    },
-    streamsByUser(root, args, context) {
-      return context.prisma
-        .user({
-          id: args.userId
-        })
-        .streams();
-    }
-  },
-  Mutation: {
-    createStream(root, args, context) {
-      return context.prisma.createStream({
-        title: args.title,
-        author: {
-          connect: { id: args.userId }
-        }
-      });
-    },
-    createUser(root, args, context) {
-      return context.prisma.createUser({ name: args.name, email: args.email });
-    }
-  },
-  User: {
-    streams(root, args, context) {
-      return context.prisma
-        .user({
-          id: root.id
-        })
-        .streams();
-    }
-  },
-  Stream: {
-    author(root, args, context) {
-      return context.prisma
-        .stream({
-          id: root.id
-        })
-        .author();
-    }
-  }
-};
+import { prisma } from "./generated/prisma-client";
+import { GraphQLServer } from "graphql-yoga";
+import { resolvers } from "./resolvers";
 
 const server = new GraphQLServer({
   typeDefs: "./server/src/schema.graphql",
@@ -57,4 +9,7 @@ const server = new GraphQLServer({
     prisma
   }
 });
-server.start(() => console.log("Server is running on http://localhost:4000"));
+
+server.start(() =>
+  console.log("🚀  Server is running on http://localhost:4000")
+);
